pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corporate-expense-openenv"
version = "1.0.0"
description = "Corporate expense approval RL environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.5",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "openai>=1.40",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "server*"]

[tool.setuptools]
py-modules = ["models", "client"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
